references/codebase/software/ais2025/Event-based-Eye-Tracking-Challenge-Solution/requirements.txt
h5py==3.11.0
imageio==2.37.0
numpy==1.26.3
matplotlib==3.7.3
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.2.3
PyYAML==6.0.2
tonic==1.4.3
tqdm==4.67.1
